Consolidate fragmented test classes; delete two dead test projects
The test suite had a long tail of one-test-per-bug-fix classes: 37 files held
1-3 tests each -- 15% of the files carrying 1.2% of the tests. This merges the
genuine fragments and removes code that was never being run.

Dead projects (never in Zork.sln, so `dotnet test` in CI has never run them):
  - Lambda/test/Lambda.Tests and Planetfall-Lambda/test/Planetfall-Lambda.Tests.
    Both ValuesControllerTests target a ValuesController that no longer exists
    in either Lambda src, so they would fail if they ever ran; Lambda.Tests'
    GameRequestTests is fully superseded by UnitTests/Models/GameRequestTests.
    They were also the only xUnit in an otherwise all-NUnit repo. Lambda
    coverage lives in UnitTests/Lambda and UnitTests/PlanetfallLambda.
  - UnitTests/ConversationPatternTests.cs: 3 bytes, a bare UTF-8 BOM.

Merges (ZorkOne.Tests/Places 20 -> 11 files):
  UndergroundPassagesTests, HighPlacesTests, WaterTests, EchoAndMagicWordTests.
  The last one merges three files that already cross-referenced each other --
  the two #354 free-command-word cases are deliberate mirrors.

Merges (UnitTests):
  MetaCommandTests, TakeCommandTests, ContainerTests,
  ReadAndExamineProcessorTests, SimpleVerbProcessorTests, and DisambiguationTests
  -- disambiguation was being tested from three separate directories (processor
  matching rules, end-to-end flow, save/restore persistence), which made it easy
  to fix one layer and miss the others.

Merges (Planetfall.Tests): DoorTests, CorridorTransitionTests.

Two dedups beyond file count: four Places fixtures each carried a private copy
of the same lit-lantern setup, and seven global-command fixtures repeated the
same real-IntentParser construction in every test. Both are now single helpers
on EngineTestsBase (GetLitTargetAt<T> and GetTargetWithRealParser).

Also renamed for accuracy: TakeSwordRepro (folded into TakeCommandTests),
NumberConverter -> NumberConverterTests, and a look test that was living under
an inventory-sounding name.

No behavior changes and no coverage lost -- every merge was a verbatim move of
test bodies, with renames only where merging caused a method-name collision.
Test counts are identical per project before and after (UnitTests 1095,
ZorkOne.Tests 1787, Planetfall.Tests 3227, EscapeRoom.Tests 9); all suites pass.

Not done here: Planetfall.Tests still has ~57 files in a flat directory while
ZorkOne.Tests has Places/People/Things. Its existing Hints/ and Walkthrough/
subfolders use folder-matching namespaces, so foldering it properly is a
namespace change across every file -- a large rename sweep, worth its own PR.
