+++ Lambda/src/Lambda/Readme.md
@@ -70,13 +70,6 @@ If already installed check if new version is available.
     dotnet tool update -g Amazon.Lambda.Tools
 ```
 
-Execute unit tests
-
-```
-    cd "Lambda/test/Lambda.Tests"
-    dotnet test
-```
-
 Deploy application
 
 ```
