+++ UnitTests/OpenAIClientBehaviorTests.cs
@@ -86,6 +86,30 @@ public class ChatGPTClientBehaviorTests
             It.IsAny<ChatCompletionOptions>()), Times.Once);
     }
 
+    // NormalizeQuotes is the client's output-sanitizing step: the model likes to emit typographic
+    // quotes, which then leak into a game whose text is otherwise plain ASCII. Merged in from the
+    // former standalone ChatGPTClientTests.
+    [Test]
+    public void NormalizeQuotes_ConvertsCurlyDoubleQuotes_ToStraight()
+    {
+        var input = "Floyd points. “This one is broken.”";
+        ChatGPTClient.NormalizeQuotes(input).Should().Be("Floyd points. \"This one is broken.\"");
+    }
+
+    [Test]
+    public void NormalizeQuotes_ConvertsCurlyApostrophes_ToStraight()
+    {
+        ChatGPTClient.NormalizeQuotes("Floyd’s gears feel creaky.")
+            .Should().Be("Floyd's gears feel creaky.");
+    }
+
+    [Test]
+    public void NormalizeQuotes_LeavesStraightQuotesUntouched()
+    {
+        const string already = "Floyd asks, \"What's that?\"";
+        ChatGPTClient.NormalizeQuotes(already).Should().Be(already);
+    }
+
     private static Mock<IChatCompletionClient> CompletionReturning(string response)
     {
         var completion = new Mock<IChatCompletionClient>();
