--- UnitTests/SingleNounProcessors/EatProcessorTests.cs
@@ -1,93 +1,0 @@
-using GameEngine;
-using GameEngine.Item.ItemProcessor;
-using Model.AIGeneration;
-using Model.AIGeneration.Requests;
-using Model.Intent;
-using Model.Interface;
-
-namespace UnitTests.SingleNounProcessors;
-
-public class EatProcessorTests : EngineTestsBase
-{
-    [Test]
-    public async Task Eat()
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<Kitchen>();
-
-        // Act
-        await target.GetResponse("open sack");
-        await target.GetResponse("take lunch");
-        var result = await target.GetResponse("eat lunch");
-
-        // Assert
-        result.Should().Contain("hit the spot");
-    }
-
-    [Test]
-    public async Task Eat_BeforeTakingIt()
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<Kitchen>();
-
-        // Act
-        await target.GetResponse("open sack");
-        var result = await target.GetResponse("eat lunch");
-
-        // Assert
-        result.Should().Contain("hit the spot");
-        result.Should().Contain("Taken");
-    }
-
-    [Test]
-    public async Task Eat_SecondTime()
-    {
-        var target = GetTarget();
-        Client.Setup(s => s.GenerateNarration(It.IsAny<NounNotPresentOperationRequest>(), It.IsAny<string>()))
-            .ReturnsAsync("BOB");
-        target.Context.CurrentLocation = Repository.GetLocation<Kitchen>();
-
-        // Act
-        await target.GetResponse("take sack");
-        await target.GetResponse("open sack");
-        await target.GetResponse("take lunch");
-        await target.GetResponse("eat lunch");
-        var result = await target.GetResponse("eat lunch");
-
-        // Assert
-        result.Should().Contain("BOB");
-    }
-
-    [Test]
-    public async Task EatProcessor_WrongVerb()
-    {
-        var target = GetTarget();
-        Client.Setup(s => s.GenerateNarration(It.IsAny<VerbHasNoEffectOperationRequest>(), It.IsAny<string>()))
-            .ReturnsAsync("BOB");
-        target.Context.CurrentLocation = Repository.GetLocation<Kitchen>();
-
-        // Act
-        await target.GetResponse("open sack");
-        await target.GetResponse("take lunch");
-        var result = await target.GetResponse("throw lunch");
-
-        // Assert
-        result.Should().Contain("BOB");
-    }
-
-    [Test]
-    public void EatProcessor_WrongType()
-    {
-        IVerbProcessor target = new EatAndDrinkInteractionProcessor();
-        target.Process(Mock.Of<SimpleIntent>(), Mock.Of<IContext>(), new Lantern(), Mock.Of<IGenerationClient>())
-            .Result.Should().BeNull();
-    }
-
-    [Test]
-    public void EatProcessor_WrongBaseType()
-    {
-        IVerbProcessor target = new EatAndDrinkInteractionProcessor();
-        Assert.Throws<Exception>(() => target.Process(Mock.Of<SimpleIntent>(), Mock.Of<IContext>(), new MadeUpItem(),
-            Mock.Of<IGenerationClient>()));
-    }
-}
