+++ UnitTests/SingleNounProcessors/ReadAndExamineProcessorTests.cs
@@ -2,8 +2,38 @@ using GameEngine;
 
 namespace UnitTests.SingleNounProcessors;
 
-public class ExamineProcessorTests : EngineTestsBase
+/// <summary>
+/// Read and examine are the two "look closely at a thing" verbs, and they share every interesting
+/// edge: the same leaflet text, the same container-state reporting, and the same darkness guard.
+/// Merged from the former ReadProcessorTest / ExamineProcessorTests.
+/// </summary>
+public class ReadAndExamineProcessorTests : EngineTestsBase
 {
+    [Test]
+    public async Task ReadProcessor()
+    {
+        var target = GetTarget();
+
+        // Act
+        await target.GetResponse("open mailbox");
+        var result = await target.GetResponse("read leaflet");
+
+        result.Should().Contain("low cunning");
+    }
+
+    [Test]
+    public async Task ReadInTheDarkProcessor()
+    {
+        var target = GetTarget();
+        target.Context.CurrentLocation = Repository.GetLocation<Attic>();
+        target.Context.Take(Repository.GetItem<Leaflet>());
+
+        // Act
+        var result = await target.GetResponse("read leaflet");
+
+        result.Should().Contain("too dark");
+    }
+
     [Test]
     public async Task ExamineProcessor()
     {
