--- ZorkOne.Tests/Places/LoudRoomTests.cs
@@ -1,101 +1,0 @@
-﻿using FluentAssertions;
-using GameEngine;
-using ZorkOne.Item;
-using ZorkOne.Location;
-
-namespace ZorkOne.Tests.Places;
-
-[TestFixture]
-public class LoudRoomTests : EngineTestsBase
-{
-    [Test]
-    [TestCase("W")]
-    [TestCase("walk w")]
-    [TestCase("go west")]
-    [TestCase("head 'west'")]
-    [TestCase("move \"west\"")]
-    public async Task CanLeave(string command)
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<LoudRoom>();
-        target.Context.ItemPlacedHere(Repository.GetItem<Torch>());
-
-        var response = await target.GetResponse(command);
-        Console.WriteLine(response);
-
-        response.Should().Contain("Round Room");
-    }
-
-    [Test]
-    public async Task EchosLastWordsByDefault()
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<LoudRoom>();
-
-        var response = await target.GetResponse("hi");
-        Console.WriteLine(response);
-
-        response.Should().Contain("hi hi ...");
-    }
-
-    [Test]
-    public async Task EchoingAFreeCommandWord_StillCountsAsATurn()
-    {
-        // Issue #354 follow-up: the room's echo catch-all fires for ANY non-direction word,
-        // including ones that happen to be global free-command names ("score"). Since the room -
-        // not ScoreProcessor - actually produced the response, it must still count as a real turn
-        // (Moves increments), exactly like the "hi" control case above.
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<LoudRoom>();
-
-        var response = await target.GetResponse("score");
-        Console.WriteLine(response);
-
-        response.Should().Contain("score score ...");
-        target.Moves.Should().Be(1);
-    }
-
-    [Test]
-    public async Task CanStillQuit()
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<LoudRoom>();
-
-        var response = await target.GetResponse("quit");
-        Console.WriteLine(response);
-
-        response.Should().Contain("Do you wish");
-    }
-
-    [Test]
-    public async Task EchosLastWorkByDefault_MultiPhrase()
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<LoudRoom>();
-
-        var response = await target.GetResponse("hi there buddy");
-        Console.WriteLine(response);
-
-        response.Should().Contain("buddy buddy ...");
-    }
-
-    [Test]
-    [TestCase("echo")]
-    [TestCase("\"echo\"")]
-    [TestCase("'echo'")]
-    [TestCase("say the word 'echo'")]
-    [TestCase("shout echo")]
-    [TestCase("yell the word \"echo\"")]
-    [TestCase("scream 'echo'")]
-    public async Task SimplestSolution(string input)
-    {
-        var target = GetTarget();
-        target.Context.CurrentLocation = Repository.GetLocation<LoudRoom>();
-
-        var response = await target.GetResponse(input);
-        Console.WriteLine(response);
-
-        response.Should().Contain("The acoustics of the room");
-        Repository.GetLocation<LoudRoom>().EchoHasBeenSpoken.Should().BeTrue();
-    }
-}
